Fix: Larivaar unicode break (#537)
* Fix: Larivaar unicode break

* use ImprovedLarivaar
